Add getters and setters for height and color options
Allow updating the `waveColor`, `progressColor`, `cursorColor` and
`height` options that were passed as part of the options for create.

For the wave color options, simply ensure that the wave is
redrawn. For the height option, delegate to `drawer.setHeight` to
handle the update.

The cursor is a border on one of the wave elements. Add an
`updateCursor` method to the drawer interface to allow updating the
cursor related styles.